Fix minor bugs in tests
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@3403 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/samson/common/Rate.cpp
@@ -29,12 +29,12 @@ namespace samson {
         return rate_kvs.getTotalSize();
     }
     
-    size_t Rate::get_rate_size()
+    double Rate::get_rate_size()
     {
         return rate_size.getRate();
     }
     
-    size_t Rate::get_rate_kvs()
+    double Rate::get_rate_kvs()
     {
         return rate_kvs.getRate();
     }
